+++ src/main/java/com/CuteNekoDragon/Core/common/item/SVOTiers.java
@@ -1,6 +1,5 @@
 package com.CuteNekoDragon.Core.common.item;
 
-import com.CuteNekoDragon.Core.SVOCore;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.world.item.Items;
@@ -9,6 +8,8 @@ import net.minecraft.world.item.Tiers;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraftforge.common.ForgeTier;
 import net.minecraftforge.common.TierSortingRegistry;
+
+import com.CuteNekoDragon.Core.SVOCore;
 
 import java.util.List;
 
@@ -23,10 +24,8 @@ public class SVOTiers {
                     1.5f,
                     14,
                     BlockTags.MINEABLE_WITH_PICKAXE,
-                    () -> Ingredient.of(Items.COPPER_INGOT)
-            ),
+                    () -> Ingredient.of(Items.COPPER_INGOT)),
             new ResourceLocation(SVOCore.MOD_ID, "copper"),
             List.of(Tiers.STONE),
-            List.of(Tiers.IRON)
-    );
+            List.of(Tiers.IRON));
 }